This PR was merged into the 1.9-dev branch.

Discussion
----------

| Q               | A
| --------------- | -----
| Branch?         | master
| Bug fix?        | yes
| New feature?    | no
| BC breaks?      | no
| Deprecations?   | no
| License         | MIT

<!--
 - Bug fixes must be submitted against the 1.7 or 1.8 branch (the lowest possible)
 - Features and deprecations must be submitted against the master branch
 - Make sure that the correct base branch is set

 To be sure you are not breaking any Backward Compatibilities, check the documentation:
 https://docs.sylius.com/en/latest/book/organization/backward-compatibility-promise.html
-->


Commits
-------

b0fcc2cb3891fd9a2e69d80529eca2143a763426 Upgrade psalm and add fixes

+++ src/Sylius/Component/Core/Customer/Statistics/CustomerStatisticsProvider.php
@@ -64,7 +64,7 @@ final class CustomerStatisticsProvider implements CustomerStatisticsProviderInte
      */
     private function getOrdersSummedTotal(array $orders): int
     {
-        return (int) array_sum(
+        return array_sum(
             array_map(function (OrderInterface $order) {
                 return $order->getTotal();
             }, $orders)
